Retry: migrate tests to munit
